Split js files into relevant folders
