+++ src/components/about/study-duration.js
